Move client test out of BPF build
