+++ themis/client/Cargo.toml
@@ -1,10 +1,8 @@
 
-# Note: This crate must be built using do.sh
-
 [package]
-name = "spl-themis"
+name = "spl-themis-client"
 version = "0.1.0"
-description = "Solana Program Library THEMIS"
+description = "SPL THEMIS client"
 authors = ["Solana Maintainers <maintainers@solana.foundation>"]
 repository = "https://github.com/solana-labs/solana-program-library"
 license = "Apache-2.0"
@@ -18,16 +16,16 @@ program = ["solana-sdk/program"]
 default = ["solana-sdk/default"]
 
 [dependencies]
-bincode = "1.3"
+
+[dev-dependencies]
 curve25519-dalek = { version = "2.0.0", default-features = false, features = ["serde"] }
 elgamal_ristretto = { git = "https://github.com/iquerejeta/elgamal", rev = "05e5481e5b7c5b596963ebe0465aab7879cd0b8b" }
-getrandom = { version = "0.1.14", features = ["dummy"] }
-num-derive = "0.3"
-num-traits = "0.2"
-rand = "0.7.0"
-serde = "1.0"
-solana-sdk = { version = "1.3.11", default-features = false, optional = true }
-thiserror = "1.0"
+solana-banks-client = "1.3.11"
+solana-banks-server = "1.3.11"
+solana-runtime = "1.3.11"
+solana-sdk = "1.3.11"
+spl-themis = { version = "0.1.0", path = "../program" }
+tokio = "0.2"
 
 [lib]
 crate-type = ["cdylib", "lib"]
